feat(platform): compile cleanly on macOS (platform-compile)
First phase of the macOS port: make the Rust backend compile for
aarch64-apple-darwin without changing Linux behavior.

- Cargo.toml: scope libc to cfg(unix); split keyring per target
  (async-secret-service/Linux, apple-native Keychain/macOS, no plaintext
  fallback); gate zbus to Linux.
- updater.rs: gate the D-Bus FileManager1 reveal to Linux, non-Linux
  reveal is a log-only no-op (opener plugin lands in platform-desktop).
- worktree_sessions.rs: widen statvfs f_bavail/f_frsize to u64 (Darwin
  types them u32/u64; Linux both u64) so the multiply typechecks.
- commands.rs: replace Command::new("which") with the which crate.
- shim.rs: scope find_in_environ to its real call sites.
- opencode adapter: keep XDG paths on macOS (verified vs dirs::config_dir,
  which would point at the wrong dir) — comment only.
- CI: new macos-cross-check job (cargo check --target, macos-latest runner;
  a Linux runner cannot build the objc2 stack). CLAUDE.md cross-platform
  invariant.

Also lands the OpenSpec multiplatform port plan: 5 changes
(platform-compile implemented; proc/desktop/ipc/bundle-ci planned).

macOS cross-check clean; Linux clippy -D warnings + 684 tests + fmt + tsc green.

+++ src-tauri/src/mcp/worktree_sessions.rs
@@ -592,6 +592,10 @@ pub struct GateResourceInfo {
     pub over_soft_cap: bool,
 }
 
+// The `as u64` widening is a no-op on Linux (both fields are already u64) but
+// required on macOS, where Darwin types `f_bavail` as u32 and `f_frsize` as
+// u64 — clippy can only see the Linux widths, so it wrongly flags the cast.
+#[allow(clippy::unnecessary_cast)]
 fn free_disk_bytes(path: &std::path::Path) -> u64 {
     // statvfs: f_bavail (blocks available to non-root) * f_frsize.
     use std::os::unix::ffi::OsStrExt;
@@ -603,7 +607,10 @@ fn free_disk_bytes(path: &std::path::Path) -> u64 {
     unsafe {
         let mut stat: libc::statvfs = std::mem::zeroed();
         if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-            stat.f_bavail.saturating_mul(stat.f_frsize)
+            // Darwin types these narrower than Linux (f_bavail: u32, f_frsize:
+            // u64 on macOS; both u64 on Linux) — widen both to u64 so the
+            // multiply typechecks and never truncates on either target.
+            (stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64)
         } else {
             0
         }
